% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_languages}
\alias{get_available_languages}
\title{List the available languages that can be used to highlight}
\usage{
get_available_languages()
}
\description{
List the available languages that can be used to highlight
}
