% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_themes}
\alias{get_available_themes}
\title{List the available themes that can be used with highlighter}
\usage{
get_available_themes()
}
\description{
List the available themes that can be used with highlighter
}
