% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlight}
\alias{highlight}
\title{Highlight any R code within htmlwidget.}
\usage{
highlight(code, width = "100\%", height = "100\%")
}
\arguments{
\item{code}{Code you want to highlight.}
}
\description{
Highlight any R code within htmlwidget.
}
\examples{
highlight("a <- 1:10")

}

