% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlighter}
\alias{highlighter}
\title{Highlighter}
\usage{
highlighter(
  code,
  language = "r",
  plugins = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{language}{The programming language chosen to be highlighted}

\item{plugins}{Optional. A list of plugins to be used}
}
\description{
Highlights code
}
