% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{line_number}
\alias{line_number}
\title{Line Number plugin}
\usage{
line_number(use_line_number = TRUE, start_from = 1)
}
\arguments{
\item{use_line_number}{Logical}

\item{start_from}{A numeric indicating where to start to count from}
}
\description{
Line Number plugin
}
