% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_highlighter_theme}
\alias{set_highlighter_theme}
\title{Set Highlighter Theme
Sets the highlighter theme globally}
\usage{
set_highlighter_theme(theme)
}
\arguments{
\item{theme}{A character. Indicating which theme will be used}
}
\description{
Set Highlighter Theme
Sets the highlighter theme globally
}
