% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_languages}
\alias{get_available_languages}
\title{Lists the current available Languages}
\usage{
get_available_languages()
}
\value{
A character vector that contains the programming languages available
to highlight.
}
\description{
List the available languages that can be used to highlight
}
