% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{highlight}
\alias{highlight}
\title{Line highlight plugin}
\usage{
highlight(range)
}
\arguments{
\item{range}{A character indicating the range to be used, for example 2-5 will highlight from 2
up to 5. Also you can highlight two or more ranges in the following way 2-5,10-13,19}
}
\description{
Line highlight plugin
}
