% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{highlight}
\alias{highlight}
\title{Line highlight plugin}
\usage{
highlight(range)
}
\arguments{
\item{range}{A character indicating the range to be used, for example 2-5
will highlight from 2 up to 5. Also you can highlight two or more ranges in
the following way 2-5,10-13,19.}
}
\value{
A list (named) with the name of the plugin and the range passed
by the function
}
\description{
Line highlight plugin
}
\examples{
if (interactive()) {
  highlighter::highlighter(
    "print('Hello, world!')\ncat <- \"Aristofanes\"\nstr(some_variable)",
    language = "r",
    plugins = list(
      highlight(
        range = "1-2"
      )
    )
  )
}

}
