% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Highlight code}
\usage{
highlight(
  code = NULL,
  file = NULL,
  language = "r",
  style = "default",
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{code}{The original raw code.}

\item{file}{The file want highlight.}

\item{language}{Language name, supports r, python and sql. The default is r.}

\item{style}{Highlight style. The default is default.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended. The default is \code{'100\%'}.}

\item{elementId}{Use an explicit element ID for the widget (rather than an automatically generated one). The default is NULL.}
}
\description{
Highlight code using highlight.js
}
\examples{
highlight(code = "df <- head(mtcars)")
highlight(code = "df <- head(mtcars)", style = "dark")

}
