% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlight_file}
\alias{highlight_file}
\title{Highlights the content of a given file}
\usage{
highlight_file(file_path)
}
\arguments{
\item{file_path}{The path to the file to be highlighted}
}
\description{
Highlights the content of a given file
}
