% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlight_file}
\alias{highlight_file}
\title{Highlight Syntax of a File}
\usage{
highlight_file(file_path, language = NULL, plugins = NULL, theme = "default")
}
\arguments{
\item{file_path}{The path to the file to be highlighted}

\item{language}{The programming language chosen to be highlighted}

\item{plugins}{Optional. A list of plugins to be used}

\item{theme}{A character. Indicating which theme will be used}
}
\value{
An object of class \code{highlighter}
}
\description{
Highlights the content of a given file according to the source
language, theme and plugins used.
}
\seealso{
\code{\link[=get_available_languages]{get_available_languages()}} for available languages,
\code{\link[=get_available_themes]{get_available_themes()}} for available themes
}
