% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlighter}
\alias{highlighter}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
highlighter(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
