% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlighter}
\alias{highlighter}
\title{Highlighter}
\usage{
highlighter(
  code,
  language = "r",
  plugins = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{code}{The code to be highlighted}

\item{language}{The programming language chosen to be highlighted}

\item{plugins}{Optional. A list of plugins to be used}

\item{width}{Optional. The width to be used by the widget}

\item{height}{Optional. The height to be used by the widget}

\item{elementId}{Optional. The DOM element id to be used by the widget}
}
\description{
Highlights code
}
